@@ -1,9 +1,18 @@
 class TasksController < ApplicationController
-  #:autenticacao, :except => [:index, :show] # linha adicionada
   authorize_resource :except => [:index]
   
   layout 'application', :except => :console
     
+
+  def carrega_agregados
+    @authors = Author.all
+    @categories = Category.all
+    @sites = Site.all
+    @criticidades = Criticidade.all
+    @fornecedores = Fornecedor.all
+    @status_incidentes = StatusIncidente.all
+  end
+
   def index
 
     @search_tasks = params[:search] || session[:search_tasks] 
@@ -18,9 +27,12 @@ class TasksController < ApplicationController
         @tasks = Task.search @search_tasks, :match_mode => :boolean #, :per_page => 20, :page => params[:page]
         @tasks.length
         @tasks.compact!
+        
       else 
          puts "ativos"
-         @tasks = Task.ativos
+         @tasks = Task.abertas
+         @tasksativas = Task.ativas_nao_abertas
+
       end
     else
          puts "publicas"
@@ -51,44 +63,35 @@ class TasksController < ApplicationController
     @ci = Ci.find(params[:id])
     puts @ci
     @task = Task.new
-    @authors = Author.all
-    @categories = Category.all
-    @sites = Site.all
-    @criticidades = Criticidade.all
-    @fornecedores = Fornecedor.all
+    carrega_agregados
+   
     # new
     @task.site_id = @ci.site_id
     @task.ci_id = @ci.id
     @task.nome = @ci.descricao
     @task.tipotask = "Incidente"
     @task.fornecedor_id = @ci.contrato.fornecedor_id if @ci.contrato
+    @task.solicitante = @ci.Owner
     render :new
   end
 
   def new
     @task = Task.new
-    @authors = Author.all
-    @categories = Category.all
-    @sites = Site.all
-    @criticidades = Criticidade.all
-    @fornecedores = Fornecedor.all
-    #@tipotasks = Tipotask.all
+    carrega_agregados
   end
 
   def edit
     @task = Task.find(params[:id])
-    @authors = Author.all
-    @categories = Category.all
-    @sites = Site.all
-    @criticidades = Criticidade.all
-    @fornecedores = Fornecedor.all
+    carrega_agregados
     #@tipotasks = Tipotask.all
   end
   
-  def distribui(subject,task)
-    #message = UserMailer.email_alerta("josecarlosmagno@me.com",subject,task)
-    #puts message
-    #message.deliver
+  def distribui(task)
+    p = Hash[:alerta => task.id]
+    #TODO - mudar o 3 para um search (alerta, subtipo)
+    job = JobEnviarEmail.criar(3, p.to_yaml)
+    EnviaEmailWorker.perform_async(job.id)
+    # usar o sidekiq para distribuir email..
     
   end    
 
@@ -97,7 +100,7 @@ class TasksController < ApplicationController
     @task.Ativo = true;
     respond_to do |format|
       if @task.save
-        distribui("Alerta: #{@task.status}",@task)
+        distribui(@task)
         format.html { redirect_to(@task, :notice => 'Alerta criado com sucesso.') }
       else
         flash[:error] = "<ul>" + @task.errors.full_messages.map{|o| "<li>" + o + "</li>" }.join("") + "</ul>"
@@ -118,7 +121,7 @@ class TasksController < ApplicationController
     @task = Task.find(params[:id])
     respond_to do |format|
       if @task.update_attributes(params[:task])
-        distribui("Alerta: #{@task.status} - Alteracao",@task) 
+        #distribui("Alerta: #{@task.status} - Alteracao",@task) 
         format.html { redirect_to(@task, :notice => 'Tarefa alterada com sucesso.') }
       else
         flash[:error] = "<ul>" + @task.errors.full_messages.map{|o| "<li>" + o + "</li>" }.join("") + "</ul>"
